Refine receptor validation and cache reference data
